source code
GE423 Mechatronics Project, University of Illinois at Urbana Champaign

The Mobile Robot uses an onboard LADAR to map the area in 2D
instantaneously and travels between checkpoints in the maze by path
planning using A*. Kalman filtering is applied to dead reckoned optical
encoder values and values from a motion capture system to get accurate
positioning of the robot. Computer Vision is used with an onboard
camera and a servo mechanism to detect, collect, sort and deposit balls
from within the maze to the drop off points.